A couple minor fixes

+++ dbg_win.c
@@ -179,7 +179,7 @@ int dbg_get_report_size(void)
 int dbg_dap_cmd(uint8_t *data, int size, int rsize)
 {
   char cmd = data[0];
-  unsigned long res;
+  DWORD res;
 
   memset(hid_buffer, 0xff, report_size + 1);
